+++ ghost/core/package.json
@@ -88,7 +88,6 @@
     "@tryghost/http-cache-utils": "0.1.17",
     "@tryghost/i18n": "0.0.0",
     "@tryghost/image-transform": "1.3.0",
-    "@tryghost/in-memory-repository": "0.0.0",
     "@tryghost/job-manager": "0.0.0",
     "@tryghost/kg-card-factory": "5.1.0",
     "@tryghost/kg-converters": "1.1.0",
