--- ghost/in-memory-repository/src/index.ts
@@ -1,1 +1,0 @@
-export {InMemoryRepository} from './InMemoryRepository';
