--- ghost/in-memory-repository/src/nql.d.ts
@@ -1,1 +1,0 @@
-declare module '@tryghost/nql'
